% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitLinearModel.R
\name{fitLinearModels}
\alias{fitLinearModels}
\title{Fit a series of linear model}
\usage{
fitLinearModels(x, y, grades, newData)
}
\arguments{
\item{grades:}{the orders of the fitted models}

\item{newData:}{a vector containing the x value where the fitted model sholud be evalued.}

\item{x:}{the indipendent variable of the model}

\item{y:}{the dipendent variable of the model, ie the observations}
}
\value{
a vector containing the vectors of the predicted values.
}
\description{
This function fits a linear model for every grade from "lowerGrade" to "upperGrade".
It returns the predicted values in a given number of points, from the minimum to
the maximum of the x values passed.
}

